@@ -1,13 +1,11 @@
 ﻿@page "/devices"
 
 @rendermode InteractiveServer
-@inject AppDbContext DbContext
+@inject IDbContextFactory<AppDbContext> DbContextFactory
 @inject TorizonService TorizonService
 @inject DeviceService DeviceService
 
 <PageTitle>Dispositivos</PageTitle>
-
-<WindowDimensions OnChange="OnDimensionsChanged" />
 
 @if (isLoading)
 {
@@ -22,138 +20,207 @@
 }
 else
 {
-    <a href="/home" class="text-black">
-        <i class="bi bi-arrow-left fs-1"></i>
-    </a>
-
-    <br />
-    <br />
-
-    <div class="w-50">
-        <input type="text" class="form-control" placeholder="Pesquisar dispositivo pelo nome..."
-               @bind="searchTerm" @bind:event="oninput" />
-    </div>
-
-    <br />
-
-    <div class="d-flex justify-content-between align-items-center mt-2">
-        <div class="d-flex gap-4">
-
-            <div class="d-flex flex-column">
-                <label for="nameOrder" style="font-size: 18px; font-weight: bold;">Nome</label>
-                <select id="nameOrder" class="form-select" @onchange='e => OnOrderChanged(e, "DeviceName")'>
-                    <option value="relevance">Relevância</option>
-                    <option value="ascendent">Ascendente</option>
-                    <option value="descendent">Descendente</option>
-                </select>
-            </div>
-
-            <div class="d-flex flex-column">
-                <label for="hibernationOrder" style="font-size: 18px; font-weight: bold;">Hibernação</label>
-                <select id="hibernationOrder" class="form-select" @onchange='e => OnOrderChanged(e, "Hibernation")'>
-                    <option value="everything">Tudo</option>
-                    <option value="yes">Sim</option>
-                    <option value="no">Não</option>
-                </select>
-            </div>
-
-            <div class="d-flex flex-column">
-                <label for="hourOrder" style="font-size: 18px; font-weight: bold;">Data de Criação</label>
-                <select id="hourOrder" class="form-select" @onchange='e => OnOrderChanged(e, "CreatedAt")'>
-                    <option value="relevance">Relevância</option>
-                    <option value="ascendent">Ascendente</option>
-                    <option value="descendent">Descendente</option>
-                </select>
-            </div>
-
-            <div class="d-flex flex-column">
-                <label for="statusOrder" style="font-size: 18px; font-weight: bold;">Estado</label>
-                <select id="statusOrder" class="form-select" @onchange='e => OnOrderChanged(e, "Status")'>
-                    <option value="all">Tudo</option>
-                    <option value="uptodate">UpToDate</option>
-                    <option value="error">Error</option>
-                    <option value="notseen">NotSeen</option>
-                    <option value="updatepending">UpdatePending</option>
-                    <option value="updatescheduled">UpdateScheduled</option>
-                    <option value="outdated">Outdated</option>
-                </select>
-            </div>
-
+    <div class="custom-container d-flex flex-column">
+        <div class="page-title">
+            <i class="bi bi-card-list fs-3" style="padding-left: 10px;"></i>
+            <h3 class="pt-2">Dispositivos</h3>
         </div>
 
-        <div class="d-flex justify-content-between align-items-center gap-3">
-            <span>@actualPage - @TotalPages de @TotalPages</span>
-            <div class="d-flex gap-2">
-                <button class="btn custom-outline-button btn-lg"
-                        @onclick="PrevPage" disabled="@(!CanGoPrev)">
-                    &lt;
-                </button>
+        <div class="d-flex flex-column flex-md-row justify-content-between align-items-start align-items-md-center mt-4 gap-3">
+            <div class="position-relative w-50">
+                <input type="text"
+                       class="form-control py-2 ps-5"
+                       placeholder="Pesquisar dispositivo pelo nome..."
+                       style="border: 1px solid grey;"
+                       @bind="searchTerm"
+                       @bind:event="oninput" />
+                <i class="bi bi-search position-absolute top-50 start-0 translate-middle-y ms-3"></i>
+            </div>
 
-                <button class="btn custom-outline-button btn-lg"
-                        @onclick="NextPage" disabled="@(!CanGoNext)">
-                    &gt;
-                </button>
+            <div class="d-flex flex-grow-1 gap-3">
+
+                <div class="d-flex align-items-center" style="width: 50%;">
+                    <label for="fleetSelect" class="fw-bold mb-0" style="min-width: 60px;">Frota</label>
+                    <select id="fleetSelect" class="form-select w-100"
+                            @bind="selectedFleetId"
+                            @bind:after="OnFleetChanged"
+                            style="border: 1px solid grey;">
+                        <option value="">Selecione uma frota</option>
+                        @foreach (var fleet in fleets)
+                        {
+                            <option value="@fleet.Id">@fleet.Name</option>
+                        }
+                    </select>
+                </div>
+
+                <div class="d-flex align-items-center" style="width: 50%;">
+                    <label for="statusOrder" class="fw-bold mb-0" style="min-width: 185px;">Estado de atualização</label>
+                    <select id="statusOrder" class="form-select w-100"
+                            @onchange='e => OnOrderChanged(e, "Status")'>
+                        <option value="all">Tudo</option>
+                        <option value="uptodate">Atualizado</option>
+                        <option value="error">Erro</option>
+                        <option value="notseen">Não Visto</option>
+                        <option value="updatepending">Pendente</option>
+                        <option value="updatescheduled">Agendado</option>
+                        <option value="outdated">Desatualizado</option>
+                    </select>
+                </div>
             </div>
         </div>
+
+        <div class="table-responsive">
+            <table class="table table-custom mt-3" style="min-width: 2000px;">
+                <thead>
+                    <tr>
+                        <th style="width: 50px; border-top-left-radius: 5px; border-right: 1px solid white;"></th>
+                        <th style="width: 250px; padding-left: 20px; text-align: left; border-right: 1px solid white;">ID</th>
+                        <th @onclick="@(() => SortByColumn("DeviceName"))"
+                            style="width: 250px; text-align: center; border-right: 1px solid white; cursor: pointer;">
+                            Nome
+                            <span style="font-size: 0.9rem; margin-left: 5px; color: @(currentSortColumn == "DeviceName" ? "white" : "rgba(255,255,255,0.5)")">
+                                @(currentSortColumn == "DeviceName" ? (sortAscending ? "▲" : "▼") : "▲")
+                            </span>
+                        </th>
+                        <th @onclick="@(() => SortByColumn("Hibernated"))"
+                            style="width: 150px; text-align: center; border-right: 1px solid white; cursor: pointer;">
+                            Hibernação
+                            <span style="font-size: 0.9rem; margin-left: 5px; color: @(currentSortColumn == "Hibernated" ? "white" : "rgba(255,255,255,0.5)")">
+                                @(currentSortColumn == "Hibernated" ? (sortAscending ? "▲" : "▼") : "▲")
+                            </span>
+                        </th>
+                        <th @onclick="@(() => SortByColumn("Notes"))"
+                            style="width: 170px; text-align: center; border-right: 1px solid white; cursor: pointer;">
+                            Notas
+                            <span style="font-size: 0.9rem; margin-left: 5px; color: @(currentSortColumn == "Notes" ? "white" : "rgba(255,255,255,0.5)")">
+                                @(currentSortColumn == "Notes" ? (sortAscending ? "▲" : "▼") : "▲")
+                            </span>
+                        </th>
+                        <th @onclick="@(() => SortByColumn("CreatedAt"))"
+                            style="width: 150px; text-align: center; border-right: 1px solid white; cursor: pointer;">
+                            Data de Criação
+                            <span style="font-size: 0.9rem; margin-left: 5px; color: @(currentSortColumn == "CreatedAt" ? "white" : "rgba(255,255,255,0.5)")">
+                                @(currentSortColumn == "CreatedAt" ? (sortAscending ? "▲" : "▼") : "▲")
+                            </span>
+                        </th>
+                        <th @onclick="@(() => SortByColumn("ActivatedAt"))"
+                            style="width: 150px; text-align: center; border-right: 1px solid white; cursor: pointer;">
+                            Data de Ativação
+                            <span style="font-size: 0.9rem; margin-left: 5px; color: @(currentSortColumn == "ActivatedAt" ? "white" : "rgba(255,255,255,0.5)")">
+                                @(currentSortColumn == "ActivatedAt" ? (sortAscending ? "▲" : "▼") : "▲")
+                            </span>
+                        </th>
+                        <th @onclick="@(() => SortByColumn("LastSeen"))"
+                            style="width: 150px; text-align: center; border-right: 1px solid white; cursor: pointer;">
+                            Última Vez Visto
+                            <span style="font-size: 0.9rem; margin-left: 5px; color: @(currentSortColumn == "LastSeen" ? "white" : "rgba(255,255,255,0.5)")">
+                                @(currentSortColumn == "LastSeen" ? (sortAscending ? "▲" : "▼") : "▲")
+                            </span>
+                        </th>
+                        <th style="width: 150px; text-align: center; border-right: 1px solid white;">Estado</th>
+                        <th style="width: 180px; text-align: center; border-top-right-radius: 5px;">Estado de atualização</th>
+                    </tr>
+
+                </thead>
+                <tbody style="max-height: 350px; padding-bottom: 60px;">
+                    @if (FilteredDevices is null || !FilteredDevices.Any())
+                    {
+                        <tr>
+                            <td colspan="5" class="text-center text-muted py-4" style="font-size: 1.1rem;">
+                                Sem dados disponíveis
+                            </td>
+                        </tr>
+                    }
+                    else
+                    {
+                        @foreach (var device in FilteredDevices)
+                        {
+                            <tr>
+                                <td style="width: 50px; vertical-align: middle;">
+                                    <i @onclick="() => ToggleMenu(device.DeviceId)" class="bi bi-three-dots-vertical px-3" style="font-size: 1.2rem; cursor: pointer;"></i>
+                                    @if (openMenuDeviceId == device.DeviceId)
+                                    {
+                                        <div class="actions-dropdown">
+                                            <button @onclick="UpdateDevice" class="dropdown-item">
+                                                <i class="bi bi-arrow-clockwise"></i> Atualizar
+                                            </button>
+                                            <button @onclick="RemoteAcess" class="dropdown-item">
+                                                <i class="bi bi-broadcast"></i> Acesso remoto
+                                            </button>
+                                        </div>
+                                    }
+                                </td>
+                                <td style="width: 250px; padding-left: 20px; word-break: break-word; white-space: normal; vertical-align: middle;">
+                                    @device.DeviceId
+                                </td>
+                                <td title="@device.DeviceName" style="width: 250px; word-break: break-word; white-space: normal; vertical-align: middle;">
+                                    @device.DeviceName
+                                </td>
+                                <td style="width: 150px; text-align: center; vertical-align: middle;">@(device.Hibernated ? "Sim" : "Não")</td>
+                                <td style="width: 170px; text-align: center; word-break: break-word; white-space: normal; vertical-align: middle;">
+                                    @(string.IsNullOrEmpty(device.Notes) ? "-" : device.Notes)
+                                </td>
+                                <td style="width: 150px; text-align: center; vertical-align: middle;">@device.CreatedAt.ToString("dd/MM/yyyy HH:mm")</td>
+                                <td style="width: 150px; text-align: center; vertical-align: middle;">@(device.LastSeen?.ToString("dd/MM/yyyy HH:mm") ?? "-")</td>
+                                <td style="width: 150px; text-align: center; vertical-align: middle;">
+                                    @(device.LastSeen?.ToString("dd/MM/yyyy HH:mm") ?? "-")
+                                </td>
+                                <td style="width: 150px; text-align: center; vertical-align: middle;">
+                                    @CheckOnlineStatus(device)
+                                </td>
+                                <td style="width: 170px; text-align: center;">
+                                    @GetStatusIcon(device.DeviceStatus!)
+                                </td>
+                            </tr>
+                        }
+                    }
+                </tbody>
+            </table>
+
+            <Modal @ref="modalSelect" Title="Selecionar Utilizador" IsVerticallyCentered="true" Size="ModalSize.Large">
+                <BodyTemplate>
+                    <SelectUserModel DeviceUuid="@currentDeviceUuid" OnNextModal="NextModalAccess" />
+                </BodyTemplate>
+            </Modal>
+
+            <Modal @ref="modalAcess" Title="Criar sessão de acesso remoto" IsVerticallyCentered="true" Size="ModalSize.Large">
+                <BodyTemplate>
+                    <RemoteAcess DeviceUuid="@currentDeviceUuid" SelectedKey="selectedSshKey" OnNextModal="NextModalLink" />
+                </BodyTemplate>
+            </Modal>
+
+            <Modal @ref="modalSsh" Title="Visualizar SSH" IsVerticallyCentered="true" Size="ModalSize.Large">
+                <BodyTemplate>
+                    <SshLinkModal DeviceUuid="@currentDeviceUuid" OnCloseModal="CloseModal" />
+                </BodyTemplate>
+            </Modal>
+
+            <Modal @ref="modalUpdate" Title="Selecionar Pacote" IsVerticallyCentered="true" Size="ModalSize.Large">
+                <BodyTemplate>
+                    <UpdateImageModal DeviceUuid="@currentDeviceUuid" OnCloseModal="CloseModal"/>
+                </BodyTemplate>
+            </Modal>
+        </div>
     </div>
-
-    <table class="table custom-table mt-5">
-        <thead style="background-color: #0c6527; color: white;">
-            <tr>
-                <th>ID</th>
-                <th>Nome</th>
-                <th style="text-align: center;">Hibernação</th>
-                <th style="text-align: center;">Notas</th>
-                <th>Data de Criação</th>
-                <th style="text-align: center;">Estado</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var device in PaginatedDevices)
-            {
-                <tr>
-                    <td title="@device.DeviceId" style="word-break: break-word; white-space: normal; vertical-align: middle;">
-                        @device.DeviceId
-                    </td>
-                    <td title="@device.DeviceName" style="word-break: break-word; white-space: normal; vertical-align: middle;">
-                        @device.DeviceName
-                    </td>
-
-                    <td style="text-align: center; vertical-align: middle;">@(device.Hibernated ? "Sim" : "Não")</td>
-                    <td style="text-align: center; vertical-align: middle;">@(string.IsNullOrEmpty(device.Notes) ? "-" : device.Notes)</td>
-                    <td style="vertical-align: middle;">@device.CreatedAt.ToString("dd/MM/yyyy HH:mm")</td>
-                    <td style="text-align: center;">
-                        @GetStatusIcon(device.DeviceStatus!)
-                    </td>
-                </tr>
-            }
-        </tbody>
-    </table>
 }
 
 @code {
-    private List<DetailedDevice> originalDevices = new List<DetailedDevice>();
-    private List<DetailedDevice> devices = new List<DetailedDevice>();
-    private List<DetailedDevice> paginatedDevices = new List<DetailedDevice>();
-    private WindowDimension? dimension;
+    // Object Variables
+    private DetailedDevice currentDevice = new();
+    private SshKey selectedSshKey = new();
+    private Modal modalSelect = new();
+    private Modal modalAcess = new();
+    private Modal modalUpdate = new();
+    private Modal modalSsh = new();
+    private List<DetailedDevice> originalDevices = new();
+    private List<DetailedDevice> devices = new();
+    private List<Fleet> fleets = new();
 
-    private int actualPage = 1;
-    private int devicesPerPage = 4;
-    private int TotalPages => (int)Math.Ceiling((double)FilteredDevices.Count() / devicesPerPage);
-
-    private bool isLoading = true;
-    private bool CanGoNext => actualPage < TotalPages;
-    private bool CanGoPrev => actualPage > 1;
-
-    private IEnumerable<DetailedDevice> FilteredDevices =>
-    string.IsNullOrWhiteSpace(searchTerm)
-        ? devices
-        : devices.Where(d => d.DeviceName.Contains(searchTerm, StringComparison.OrdinalIgnoreCase));
-
-    private IEnumerable<DetailedDevice> PaginatedDevices =>
-    FilteredDevices
-        .Skip((actualPage - 1) * devicesPerPage)
-        .Take(devicesPerPage);
+    // String Variables
+    private string? openMenuDeviceId = null;
+    private string? currentSortColumn = "DeviceName";
+    private string? currentDeviceUuid;
+    private string selectedFleetId = string.Empty;
 
     private string _searchTerm = string.Empty;
     private string searchTerm
@@ -164,116 +231,200 @@ else
             if (_searchTerm != value)
             {
                 _searchTerm = value;
-                actualPage = 1;
             }
         }
     }
 
-    private void OnDimensionsChanged(WindowDimension newDimension)
-    {
-        dimension = newDimension;
-        StateHasChanged();
-    }
+    // Boolean Variables
+    private bool isLoading = true;
+    private bool sortAscending = true;
+
+    private IEnumerable<DetailedDevice> FilteredDevices =>
+    string.IsNullOrWhiteSpace(searchTerm)
+        ? devices
+        : devices.Where(d => d.DeviceName.Contains(searchTerm, StringComparison.OrdinalIgnoreCase));
 
     protected override async Task OnInitializedAsync()
     {
         var devicesUrl = "https://app.torizon.io/api/v2beta/devices?limit=200&offset=0";
+        var fleetDevicesUrl = $"https://app.torizon.io/api/v2beta/fleets/{selectedFleetId}/devices";
 
-        bool authOk = await TorizonService.AuthenticateAsync();
-        if (authOk)
+        if(await TorizonService.AuthenticateAsync())
         {
             var apiDevices = await TorizonService.GetItemsAsync<DetailedDevice>(devicesUrl) ?? new List<DetailedDevice>();
 
             originalDevices = new List<DetailedDevice>(apiDevices);
             devices = new List<DetailedDevice>(originalDevices);
 
+            using var context = DbContextFactory.CreateDbContext();
+            fleets = await context.Fleets.ToListAsync();
+
             isLoading = false;
         }
 
-        UpdatePage();
+        StateHasChanged();
     }
 
-    private void UpdatePage()
+    private async Task NextModalAccess(SshKey key)
     {
-        paginatedDevices = devices.Skip((actualPage - 1) * devicesPerPage)
-                                 .Take(devicesPerPage)
-                                 .ToList();
+        selectedSshKey = key;
+        await modalSelect.HideAsync();
+        await modalAcess.ShowAsync();
     }
 
-    private void NextPage()
+    private async Task NextModalLink()
     {
-        if (CanGoNext)
+        await modalAcess.HideAsync();
+        await modalSsh.ShowAsync();
+    }
+
+    private async Task CloseModal()
+    {
+        await modalAcess.HideAsync();
+        await modalUpdate.HideAsync();
+        await modalSsh.HideAsync(); 
+    }
+
+    private async Task UpdateDevice()
+    {
+        await modalUpdate.ShowAsync();
+        openMenuDeviceId = null;
+    }
+
+    private async Task RemoteAcess()
+    {
+        await modalSelect.ShowAsync();
+        openMenuDeviceId = null;
+    }
+
+    private void ToggleMenu(string deviceId)
+    {
+        if (openMenuDeviceId == deviceId)
         {
-            actualPage++;
-            UpdatePage();
+            openMenuDeviceId = null;
+        }
+        else
+        {
+            openMenuDeviceId = deviceId;
+            currentDevice = devices.FirstOrDefault(d => d.DeviceId == deviceId);
+            currentDeviceUuid = currentDevice?.DeviceUuid;
         }
     }
 
-    private void PrevPage()
+    private async Task OnFleetChanged()
     {
-        if (CanGoPrev)
+        if (string.IsNullOrWhiteSpace(selectedFleetId))
         {
-            actualPage--;
-            UpdatePage();
+            var allDevices = await TorizonService.GetItemsAsync<DetailedDevice>("https://app.torizon.io/api/v2beta/devices?limit=200&offset=0");
+            devices = allDevices?.ToList() ?? new();
         }
+        else
+        {
+            var url = $"https://app.torizon.io/api/v2beta/fleets/{selectedFleetId}/devices";
+            var fleetDevices = await TorizonService.GetItemsAsync<DetailedDevice>(url);
+            devices = fleetDevices?.ToList() ?? new();
+        }
+
+        StateHasChanged();
+    }
+
+    private void SortByColumn(string columnName)
+    {
+        if (currentSortColumn == columnName)
+        {
+            sortAscending = !sortAscending;
+        }
+        else
+        {
+            currentSortColumn = columnName;
+            sortAscending = true;
+        }
+
+        SortDevices();
+    }
+
+    private void SortDevices()
+    {
+        devices = currentSortColumn switch
+        {
+            "DeviceName" => (sortAscending ? devices.OrderBy(d => d.DeviceName) : devices.OrderByDescending(d => d.DeviceName)).ToList(),
+            "Hibernated" => (sortAscending ? devices.OrderBy(d => d.Hibernated) : devices.OrderByDescending(d => d.Hibernated)).ToList(),
+            "Notes" => (sortAscending ? devices.OrderBy(d => d.Notes) : devices.OrderByDescending(d => d.Notes)).ToList(),
+            "CreatedAt" => (sortAscending ? devices.OrderBy(d => d.CreatedAt) : devices.OrderByDescending(d => d.CreatedAt)).ToList(),
+            "ActivatedAt" => (sortAscending ? devices.OrderBy(d => d.ActivatedAt) : devices.OrderByDescending(d => d.ActivatedAt)).ToList(),
+            "LastSeen" => (sortAscending ? devices.OrderBy(d => d.LastSeen) : devices.OrderByDescending(d => d.LastSeen)).ToList(),
+            _ => devices
+        };
     }
 
     private MarkupString GetStatusIcon(string status)
     {
         var iconHtml = status.ToLower() switch
         {
-            "uptodate" => "<i class='bi bi-check-circle text-success fs-4'></i>",
-            "error" => "<i class='bi bi-x-circle text-danger fs-4'></i>",
-            "updatepending" => "<i class='bi bi-clock text-warning fs-4'></i>",
-            "notseen" => "<i class='bi bi-eye-slash text-muted fs-4'></i>",
-            "outdated" => "<i class='bi bi-exclamation-circle text-danger fs-4'></i>",
-            "updatescheduled" => "<i class='bi bi-calendar-check text-primary fs-4'></i>",
-            _ => "<i class='bi bi-question-circle text-secondary fs-4'></i>"
+            "uptodate" => @"<div class='state'>
+                               <i class='bi bi-check-circle text-success fs-4'></i>
+                               <span>Atualizado</span>
+                            </div>",
+            "error" => @"<div class='state'>
+                            <i class='bi bi-x-circle text-danger fs-4'></i>
+                            <span>Erro</span>
+                         </div>",
+            "updatepending" => @"<div class='state'>
+                                    <i class='bi bi-clock text-warning fs-4'></i>
+                                    <span>Pendente</span>
+                                 </div>",
+            "notseen" => @"<div class='state'>
+                              <i class='bi bi-eye-slash text-muted fs-4'></i>
+                              <span>Não Visto</span>
+                           </div>",
+            "outdated" => @"<div class='state'>
+                               <i class='bi bi-exclamation-circle text-danger fs-4'></i>
+                               <span>Desatualizado</span>
+                            </div>",
+            "updatescheduled" => @"<div class='state'>
+                                      <i class='bi bi-calendar-check text-primary fs-4'></i>
+                                      <span>Agendado</span>
+                                   </div>",
+            _ => @"<div class='state'>
+                      <i class='bi bi-question-circle text-secondary fs-4'></i>
+                      <span>Scheduled</span>
+                   </div>",
         };
+
+        return new MarkupString(iconHtml);
+    }
+
+    private MarkupString CheckOnlineStatus(DetailedDevice device)
+    {
+        bool isOnline = device.LastSeen.HasValue && (DateTime.UtcNow - device.LastSeen.Value.ToUniversalTime()).TotalMinutes < 5;
+        string iconHtml = string.Empty;
+
+        if (isOnline)
+        {
+            iconHtml = @"<div class='state'>
+                        <i class='bi bi-circle-fill text-success fs-4'></i>
+                        <span>Online</span>
+                     </div>";
+        }
+        else
+        {
+            iconHtml = @"<div class='state'>
+                        <i class='bi bi-x-circle text-danger fs-4'></i>
+                        <span>Offline</span>
+                     </div>";
+        }
 
         return new MarkupString(iconHtml);
     }
 
     private void OnOrderChanged(ChangeEventArgs e, string orderByField)
     {
-        var selectedOrder = e.Value?.ToString() ?? "relevance";
-        var selectedHibernationOrder = e.Value?.ToString() ?? "everything";
         var selectedStatusOrder = e.Value?.ToString() ?? "uptodate";
 
         switch (orderByField)
         {
-            case "CreatedAt":
-                devices = selectedOrder switch
-                {
-                    "ascendent" => originalDevices.OrderBy(d => d.CreatedAt).ToList(),
-                    "descendent" => originalDevices.OrderByDescending(d => d.CreatedAt).ToList(),
-                    "relevance" => originalDevices.ToList(),
-                    _ => devices
-                };
-                break;
-
-            case "DeviceName":
-                devices = selectedOrder switch
-                {
-                    "ascendent" => originalDevices.OrderBy(d => d.DeviceName).ToList(),
-                    "descendent" => originalDevices.OrderByDescending(d => d.DeviceName).ToList(),
-                    "relevance" => originalDevices.ToList(),
-                    _ => devices
-                };
-                break;
-
-            case "Hibernation":
-                devices = selectedHibernationOrder switch
-                {
-                    "yes" => originalDevices.Where(d => d.Hibernated).ToList(),
-                    "no" => originalDevices.Where(d => !d.Hibernated).ToList(),
-                    "everything" => originalDevices.ToList(),
-                    _ => devices
-                };
-                break;
-
             case "Status":
-                devices = selectedHibernationOrder switch
+                devices = selectedStatusOrder switch
                 {
                     "uptodate" => originalDevices.Where(d => d.DeviceStatus == "UpToDate").ToList(),
                     "error" => originalDevices.Where(d => d.DeviceStatus == "Error").ToList(),
@@ -287,7 +438,6 @@ else
                 break;
         }
 
-        actualPage = 1;
-        UpdatePage();
+        StateHasChanged();
     }
 }
