@@ -1,46 +1,109 @@
-﻿@using Microsoft.EntityFrameworkCore
+﻿@using BlazorBootstrap
+@using Microsoft.EntityFrameworkCore
 @using ToradexSwLoader.Data
 @using ToradexSwLoader.Services
-@inject IDbContextFactory<AppDbContext> DbContextFactory
 @inherits LayoutComponentBase
+@inject IDbContextFactory<AppDbContext> DbContextFactory
 @inject NavigationManager Navigation
 @inject LoginService LoginService
 @inject FinalProductStateService ProductState
+@inject SweetAlertService Swal
 
-<div class="page">
-
-    @if (IsCurrentPage(new[] { "/home", "/manageDevices", "/manageProducts", "/manageStacks", "/managePackages", "/manageFleets",
-        "/manageSecret", "/managePatterns" }))
+<div class="layout-container">
+    @if (!IsRootPage())
     {
-        <div class="sidebar">
-            <NavMenu />
-        </div>
-    }
-    else
-    {
-        <div></div>
-    }
-
-    <main>
-        @if (IsCurrentPage(new[] { "/" }))
-        {
-            <article class="content px-4">  
-                @Body
-            </article>
-        } else{
-             <div class="top-row px-4 d-flex justify-content-between">
-                <span>@username</span>
-                <span>@currentTime</span>
-                @if (IsCurrentPage(new[] { "/home" }))
+        <header class="topbar">
+            @if (ShouldShowSidebar())
+            {
+                @if (!sidebarVisible)
                 {
-                    <a href="" class="logout" target="">Terminar Sessão</a>
+                    <i class="bi bi-list fs-2" @onclick="SidebarIconOnClick" style="cursor: pointer;"></i>
                 }
-             </div>
-            <article class="content px-4">
-                @Body
-            </article>
+                else
+                {
+                    <i class="bi bi-list-nested fs-2" @onclick="SidebarIconOnClick" style="cursor: pointer;"></i>
+                }
+            }
+
+            <div class="container-fluid d-flex align-items-center justify-content-between">
+                <a class="navbar-brand" style="color: white; font-size: 1.5rem;" href="/home">ToradexSwLoader</a>
+                <span style="font-size: 1.5rem;">@currentTime</span>
+                <div class="user-menu-wrapper">
+                    <button @onclick="ToggleMenu" class="user-icon-navbar">
+                        <i class="bi bi-person-fill"></i>
+                    </button>
+
+                    @if (openMenu)
+                    {
+                        <div class="user-dropdown">
+                            <button @onclick="ChangePassword" class="dropdown-item">
+                                <i class="bi bi-lock mx-1"></i> Mudar Palavra-Passe
+                            </button>
+                            <button @onclick="OnShowModalClick" class="dropdown-item">
+                                <i class="bi bi-question-circle mx-1"></i> Sobre
+                            </button>
+                            <button @onclick="Logout" class="dropdown-item">
+                                <i class="bi bi-box-arrow-right mx-1"></i> Terminar Sessão
+                            </button>
+                        </div>
+                    }
+                </div>
+            </div>
+        </header>
+    }
+
+    <div class="main-container">
+        @if (ShouldShowSidebar())
+        {
+            <nav class="sidebar @(sidebarVisible ? "show" : "hide")">
+                <NavMenu />
+            </nav>
         }
-    </main>
+
+        <main class="content-area">
+            @Body
+
+            <Modal @ref="modalAbout" Title="Sobre" IsVerticallyCentered="true">
+                <BodyTemplate>
+                    <div style="width: 450px;">
+                        <div style="display: flex; justify-content: space-between; margin-bottom:10px;">
+                            <strong>Versão:</strong>
+                            <div style="text-align: right; flex-grow: 1;">2.0.0</div>
+                        </div>
+                        <div style="display: flex; justify-content: space-between;">
+                            <strong>URL Base:</strong>
+                            <div style="text-align: right; flex-grow: 1;">
+                                https://172.16.50.146:7023/
+                            </div>
+                        </div>
+                    </div>
+                </BodyTemplate>
+            </Modal>
+
+            <Modal @ref="modalPassword" Title="Alterar Palavra-Passe" IsVerticallyCentered="true" OnHidden="ResetPasswordForm">
+                <BodyTemplate>
+                    <div class="form-floating mb-3 position-relative">
+                        <input type="password" class="form-control pe-5" @bind="currentPassword" autocomplete="current-password" aria-required="true" placeholder="Palavra-passe atual" id="password" />
+                        <label for="password" class="form-label">Palavra-passe atual</label>
+                        <i class="bi bi-lock-fill position-absolute top-50 end-0 translate-middle-y me-3"></i>
+                    </div>
+                    <div class="form-floating mb-3 position-relative">
+                        <input type="password" class="form-control pe-5" @bind="newPassword" autocomplete="current-password" aria-required="true" placeholder="Nova Palavra-passe" id="password" />
+                        <label for="password" class="form-label">Nova Palavra-passe</label>
+                        <i class="bi bi-lock-fill position-absolute top-50 end-0 translate-middle-y me-3"></i>
+                    </div>
+                    <div class="form-floating mb-5 position-relative">
+                        <input type="password" class="form-control pe-5" @bind="confirmPassword" autocomplete="current-password" aria-required="true" placeholder="Confirmar Palavra-passe" id="password" />
+                        <label for="password" class="form-label">Confirmar Palavra-passe</label>
+                        <i class="bi bi-lock-fill position-absolute top-50 end-0 translate-middle-y me-3"></i>
+                    </div>
+                    <div class="text-center">
+                        <a class="w-50 btn btn-custom btn-lg btn-color" @onclick="ChangeNewPassword">Alterar</a>
+                    </div>
+                </BodyTemplate>
+            </Modal>
+        </main>
+    </div>
 </div>
 
 <div id="blazor-error-ui" data-nosnippet>
@@ -64,12 +127,22 @@
     }
 </style>
 
-
 @code {
-    private string? currentTime;
+    // Objects Variables
+    private Modal modalAbout = new();
+    private Modal modalPassword = new();
     private System.Timers.Timer? timer;
 
+    // String Variables
+    private string? currentTime;
     private string? username;
+    private string currentPassword = string.Empty;
+    private string newPassword = string.Empty;
+    private string confirmPassword = string.Empty;
+
+    // Boolean Variables
+    private bool openMenu = false;
+    private bool sidebarVisible = false;
 
     protected override Task OnInitializedAsync()
     {
@@ -84,6 +157,92 @@
         LoginService.OnLoginChanged += OnLoginChanged;
 
         return Task.CompletedTask;
+    }
+
+    private void SidebarIconOnClick()
+    {
+        sidebarVisible = !sidebarVisible;
+    }
+
+    private async Task OnShowModalClick()
+    {
+        await modalAbout.ShowAsync();
+        openMenu = !openMenu;
+    }
+
+    private async Task ChangePassword()
+    {
+        await modalPassword.ShowAsync();
+        openMenu = !openMenu;
+    }
+
+    private async Task ChangeNewPassword()
+    {
+        if (string.IsNullOrWhiteSpace(currentPassword) || string.IsNullOrWhiteSpace(newPassword) || string.IsNullOrWhiteSpace(confirmPassword))
+        {
+            await Swal.FireAsync(new SweetAlertOptions
+            {
+                Title = "Todos os campos são obrigatórios.",
+                Icon = SweetAlertIcon.Error,
+                ConfirmButtonColor = "#0B5120"
+            });
+            return;
+        }
+
+        if (newPassword != confirmPassword)
+        {
+            await Swal.FireAsync(new SweetAlertOptions
+            {
+                Title = "As novas palavras-passe não coincidem.!",
+                Icon = SweetAlertIcon.Error,
+                ConfirmButtonColor = "#0B5120"
+            });
+            return;
+        }
+
+        var success = await LoginService.ChangePasswordAsync(username!, currentPassword, newPassword);
+
+        if (success)
+        {
+            await Swal.FireAsync(new SweetAlertOptions
+            {
+                Title = "Palavra-passe alterada com sucesso!",
+                Icon = SweetAlertIcon.Success,
+                ConfirmButtonColor = "#0B5120"
+            });
+
+            ResetPasswordForm();
+
+            await modalPassword.HideAsync();
+        }
+        else
+        {
+            await Swal.FireAsync(new SweetAlertOptions
+            {
+                Title = "Erro ao alterar a palavra-passe. Verifique a palavra-passe atual.",
+                Icon = SweetAlertIcon.Error,
+                ConfirmButtonColor = "#0B5120"
+            });
+        }
+    }
+
+    private void ResetPasswordForm()
+    {
+        currentPassword = string.Empty;
+        newPassword = string.Empty;
+        confirmPassword = string.Empty;
+    }
+
+    private void Logout()
+    {
+        LoginService.Logout();
+        Navigation.NavigateTo("");
+        openMenu = !openMenu;
+    }
+
+    private void ToggleMenu()
+    {
+        openMenu = !openMenu;
     }
 
     private void OnLoginChanged()
@@ -116,15 +275,64 @@
         LoginService.OnLoginChanged -= OnLoginChanged;
     }
 
-    private bool IsCurrentPage(IEnumerable<string> pages)
+    private bool ShouldShowSidebar()
     {
         var relativePath = Navigation.ToBaseRelativePath(Navigation.Uri);
         relativePath = "/" + relativePath.TrimStart('/');
 
-        if (string.IsNullOrEmpty(relativePath) && pages.Contains("/"))
+        var pages = new[] { 
+            "/home", 
+            "/manageDevices", 
+            "/manageDevices/add", 
+            "/manageDevices/edit", 
+            "/manageProducts", 
+            "/manageProducts/add",
+            "/manageProducts/edit",
+            "/manageStacks", 
+            "/manageStacks/add", 
+            "/manageStacks/edit", 
+            "/managePackages", 
+            "/managePackages/add", 
+            "/managePackages/edit", 
+            "/manageFleets",
+            "/manageEntities", 
+            "/manageEntities/add", 
+            "/manageEntities/edit", 
+            "/manageUsers", 
+            "/manageUsers/add", 
+            "/manageUsers/edit", 
+            "/manageSecret", 
+            "/managePatterns", 
+            "/managePatterns/add", 
+            "/managePatterns/edit",
+            "/adminHistory",
+            "/devices",
+            "/maintenance"
+        };
+
+        if (string.IsNullOrEmpty(relativePath)) 
+            return false;
+
+        if (pages.Contains(relativePath, StringComparer.OrdinalIgnoreCase))
             return true;
 
-        return pages.Contains(relativePath, StringComparer.OrdinalIgnoreCase);
+        if (relativePath.Contains("/edit/", StringComparison.OrdinalIgnoreCase) ||
+        relativePath.EndsWith("/edit", StringComparison.OrdinalIgnoreCase))
+            return true;
+
+        return false;
     }
 
+    private bool IsRootPage()
+    {
+        var relativePath = Navigation.ToBaseRelativePath(Navigation.Uri);
+        return string.IsNullOrEmpty(relativePath) || relativePath == "/";
+    }
+
+    private bool IsHomePage()
+    {
+        var relativePath = Navigation.ToBaseRelativePath(Navigation.Uri);
+        relativePath = "/" + relativePath.TrimStart('/');
+        return relativePath.Equals("/home", StringComparison.OrdinalIgnoreCase);
+    }
 }
