@@ -4,6 +4,8 @@
 @inject FinalProductStateService ProductState
 @inject SweetAlertService Swal
 @inject TorizonService TorizonService
+@inject LoginService LoginService
+@inject NavigationManager Navigation
 
 <PageTitle>Página Inicial</PageTitle>
 
@@ -20,55 +22,81 @@
 }
 else
 {
-    <h3 class="pt-2">Últimos Carregamentos</h3>
+    <div class="custom-container d-flex flex-column">
+        <div class="page-title">
+            <i class="bi bi-card-list fs-3" style="padding-left: 10px;"></i>
+            <h3 class="pt-2">Últimos Carregamentos</h3>
+        </div>
 
-    <div class="d-flex justify-content-center align-items-center">
-        <table class="table mt-5 w-100">
-            <thead style="background-color: #0c6527; color: white;">
-                <tr>
-                    <th>Placa</th>
-                    <th class="text-center">Hora</th>
-                    <th class="text-center">Estado</th>
-                    <th></th>
-                </tr>
-            </thead>
-            <tbody>
-                @foreach (var finalProduct in finalProducts)
-                {
+        <div class="table-responsive">
+            <table class="table table-custom mt-3" style="min-width: 1300px; max-height: 600px;">
+                <thead>
                     <tr>
-                        <td style="vertical-align: middle;">@finalProduct.Name</td>
-                        <td class="text-center" style="vertical-align: middle;">@finalProduct.CreatedAt</td>
-                        <td style="text-align: center;">
-                            @GetStatusIcon(finalProduct.Status!)
-                        </td>
-                        <td class="text-center" style="vertical-align: middle;">
-                            @if (finalProduct.Status?.Equals("Outdated", StringComparison.OrdinalIgnoreCase) == true)
-                            {
-                                <button class="btn btn-danger btn-sm"
-                                        @onclick="@(() => AbortOperation(finalProduct))">
-                                    <i class="bi bi-x-circle"></i> Abortar
-                                </button>
-                            }
-                        </td>
+                        <th style="width: 350px; text-align: left; padding-left: 20px; border-top-left-radius: 5px;
+                                                       border-right: 1px solid white;">Placa</th>
+                        <th style="width: 180px; text-align: center; border-right: 1px solid white;">Hora</th>
+                        <th style="width: 150px; text-align: center; border-right: 1px solid white;">Estado</th>
+                        <th style="width: 150px; text-align: center; border-right: 1px solid white;">Estado de atualização</th>
+                        <th style="width: 180px; text-align: center; border-top-right-radius: 5px;"></th>
                     </tr>
-                }
-            </tbody>
-        </table>
-    </div>
-
-    <div class="text-center fixed-bottom pb-3 mt-3 text-muted" style="margin-left: 20vw;">
-        Última atualização: @ProductState.LastUpdate.ToString("dd/MM/yyyy HH:mm:ss")
-    </div>
+                </thead>
+                <tbody style="max-height: 400px;">
+                    @if (finalProducts is null || !finalProducts.Any())
+                    {
+                        <tr>
+                            <td colspan="5" class="text-center text-muted py-4" style="font-size: 1.1rem;">
+                                Sem dados disponíveis
+                            </td>
+                        </tr>
+                    }
+                    else
+                    {
+                        @foreach (var finalProduct in finalProducts)
+                        {
+                            <tr style="height: 60px;">
+                                <td style="width: 350px; text-align: left; vertical-align: middle; padding-left: 20px;">@finalProduct.Name</td>
+                                <td style="width: 180px; text-align: center; vertical-align: middle;">@finalProduct.CreatedAt</td>
+                                <td style="width: 150px; text-align: center; vertical-align: middle;">
+                                    @CheckOnlineStatus(finalProduct)
+                                </td>
+                                <td style="width: 150px; text-align: center; vertical-align: middle;">
+                                    @GetStatusIcon(finalProduct.Status!)
+                                </td>
+                                <td style="width: 180px; text-align: center; vertical-align: middle;">
+                                    @if (finalProduct.Status?.Equals("Outdated", StringComparison.OrdinalIgnoreCase) == true)
+                                    {
+                                        <button class="btn btn-danger btn-sm" style="width: 75%;"
+                                                @onclick="@(() => AbortOperation(finalProduct))">
+                                            <i class="bi bi-x-circle"></i> Abortar
+                                        </button>
+                                    }
+                                </td>
+                            </tr>
+                        }
+                    }
+                </tbody>
+            </table>
+        </div>
+        
+        <div class="sticky-footer-center">
+            Última atualização: @ProductState.LastUpdate.ToString("dd/MM/yyyy HH:mm:ss")
+        </div>
+    </div> 
 }
 
 @code {
+    // Object Variables
     private List<FinalProduct> paginatedFinalProducts = new List<FinalProduct>();
     private IReadOnlyList<FinalProduct> finalProducts = new List<FinalProduct>();
 
+    // Boolean Variables
     bool isLoading = true;
 
     protected override async Task OnInitializedAsync()
     {
+        CheckAuthentication();
+        Navigation.LocationChanged += OnLocationChanged;
+
         isLoading = true;
         await LoadProductsFromDb();
 
@@ -84,12 +112,28 @@ else
         };
     }
 
+    private void OnLocationChanged(object? sender, LocationChangedEventArgs e)
+    {
+        CheckAuthentication();
+    }
+
+    private void CheckAuthentication()
+    {
+        var currentUri = Navigation.ToBaseRelativePath(Navigation.Uri).ToLower();
+
+        if (!LoginService.IsLoggedIn && currentUri != "")
+        {
+            Navigation.NavigateTo("/", true);
+        }
+    }
+
     private async Task LoadProductsFromDb()
     {
         using var context = await DbContextFactory.CreateDbContextAsync();
         finalProducts = await context.FinalProducts
             .Where(p => p.Enabled)
             .OrderByDescending(fp => fp.CreatedAt)
+            .Take(10)
             .ToListAsync();
     }
 
@@ -105,6 +149,7 @@ else
     public void Dispose()
     {
         ProductState.OnChange -= OnProductsChanged;
+        Navigation.LocationChanged -= OnLocationChanged;
     }
 
     private async Task AbortOperation(FinalProduct productToRemove)
@@ -162,14 +207,58 @@ else
     {
         var iconHtml = status.ToLower() switch
         {
-            "uptodate" => "<i class='bi bi-check-circle text-success fs-4'></i>",
-            "error" => "<i class='bi bi-x-circle text-danger fs-4'></i>",
-            "updatepending" => "<i class='bi bi-clock text-warning fs-4'></i>",
-            "notseen" => "<i class='bi bi-eye-slash text-muted fs-4'></i>",
-            "outdated" => "<i class='bi bi-exclamation-circle text-danger fs-4'></i>",
-            "updatescheduled" => "<i class='bi bi-calendar-check text-primary fs-4'></i>",
-            _ => "<i class='bi bi-question-circle text-secondary fs-4'></i>"
+            "uptodate" => @"<div class='state'>
+                               <i class='bi bi-check-circle text-success fs-4'></i>
+                               <span>Atualizado</span>
+                            </div>",
+            "error" => @"<div class='state'>
+                            <i class='bi bi-x-circle text-danger fs-4'></i>
+                            <span>Erro</span>
+                         </div>",
+            "updatepending" => @"<div class='state'>
+                                    <i class='bi bi-clock text-warning fs-4'></i>
+                                    <span>Pendente</span>
+                                 </div>",
+            "notseen" => @"<div class='state'>
+                              <i class='bi bi-eye-slash text-muted fs-4'></i>
+                              <span>Não Visto</span>
+                           </div>",
+            "outdated" => @"<div class='state'>
+                               <i class='bi bi-exclamation-circle text-danger fs-4'></i>
+                               <span>Desatualizado</span>
+                            </div>",
+            "updatescheduled" => @"<div class='state'>
+                                      <i class='bi bi-calendar-check text-primary fs-4'></i>
+                                      <span>Agendado</span>
+                                   </div>",
+            _ => @"<div class='state'>
+                      <i class='bi bi-question-circle text-secondary fs-4'></i>
+                      <span>Scheduled</span>
+                   </div>",
         };
+
+        return new MarkupString(iconHtml);
+    }
+
+    private MarkupString CheckOnlineStatus(FinalProduct product)
+    {
+        bool isOnline = product.LastSeen.HasValue && (DateTime.UtcNow - product.LastSeen.Value.ToUniversalTime()).TotalMinutes < 5;
+        string iconHtml = string.Empty;
+
+        if (isOnline)
+        {
+            iconHtml = @"<div class='state'>
+                        <i class='bi bi-circle-fill text-success fs-4'></i>
+                        <span>Online</span>
+                     </div>";
+        }
+        else
+        {
+            iconHtml = @"<div class='state'>
+                        <i class='bi bi-x-circle text-danger fs-4'></i>
+                        <span>Offline</span>
+                     </div>";
+        }
 
         return new MarkupString(iconHtml);
     }
