@@ -1,4 +1,4 @@
-﻿@using System.Net.Http
+@using System.Net.Http
 @using System.Net.Http.Json
 @using System.ComponentModel.DataAnnotations
 @using Microsoft.AspNetCore.Components.Forms
@@ -13,6 +13,10 @@
 @using ToradexSwLoader.Models
 @using ToradexSwLoader.Services
 @using ToradexSwLoader.Data
+@using ToradexSwLoader.Components.Layout
 @using CurrieTechnologies.Razor.SweetAlert2
 @using System.Text.Json;
 @using System.Net.Http.Headers;
+@using BlazorBootstrap;
+@using Syncfusion.Blazor
+@using Syncfusion.Blazor.DropDowns;
