@@ -1,5 +1,6 @@
 using CurrieTechnologies.Razor.SweetAlert2;
 using Microsoft.EntityFrameworkCore;
+using Syncfusion.Blazor;
 using ToradexSwLoader.Components;
 using ToradexSwLoader.Data;
 using ToradexSwLoader.Services;
@@ -11,21 +12,21 @@ builder.Services.AddRazorComponents()
     .AddInteractiveServerComponents();
 
 var connectionString = builder.Configuration.GetConnectionString("DefaultConnection");
-builder.Services.AddDbContextFactory<AppDbContext>(options => 
-    options.UseMySql(connectionString, ServerVersion.AutoDetect(connectionString)
-));
+builder.Services.AddDbContextFactory<AppDbContext>(options =>
+    options.UseMySQL(connectionString)
+);
 
-// Configurar servi�os a serem utilizados
+// Configurar serviços a serem utilizados
 builder.Services.AddHttpClient<TorizonService>();
 builder.Services.AddScoped<DeviceService>();
-builder.Services.AddScoped<PackageService>();
 builder.Services.AddScoped<FleetService>();
 builder.Services.AddScoped<LoginService>();
-builder.Services.AddScoped<WindowService>();
 builder.Services.AddScoped<FilterService>();
 builder.Services.AddSingleton<FinalProductStateService>();
 builder.Services.AddHostedService<DeviceStatusUpdaterService>();
 builder.Services.AddSweetAlert2();
+builder.Services.AddBlazorBootstrap();
+builder.Services.AddSyncfusionBlazor();
 
 var app = builder.Build();
 
@@ -39,7 +40,7 @@ if (!app.Environment.IsDevelopment())
 
 app.UseHttpsRedirection();
 
-app.UseStaticFiles();
+app.MapStaticAssets();
 app.UseAntiforgery();
 
 app.MapRazorComponents<App>()
