@@ -1,4 +1,6 @@
-﻿using System.Net.Http.Headers;
+﻿using System.Net;
+using System.Net.Http.Headers;
+using System.Text;
 using System.Text.Json;
 using ToradexSwLoader.Models;
 
@@ -65,20 +67,102 @@ namespace ToradexSwLoader.Services
 
             return null;
         }
+
         public async Task<T?> GetItemAsync<T>(string url)
         {
-            var response = await _httpClient.GetAsync(url);
-            if (!response.IsSuccessStatusCode)
-                return default;
+            Console.WriteLine($"[HTTP] GET {url}");
 
-            var json = await response.Content.ReadAsStringAsync();
+            var response = await _httpClient.GetAsync(url);
+
+            var content = await response.Content.ReadAsStringAsync();
+            Console.WriteLine($"[HTTP] Status: {response.StatusCode}");
+            Console.WriteLine($"[HTTP] Body: {content}");
+
+            if (!response.IsSuccessStatusCode)
+            {
+                Console.WriteLine("[HTTP] Request falhou.");
+                return default;
+            }
 
             var options = new JsonSerializerOptions
             {
                 PropertyNameCaseInsensitive = true
             };
 
-            return JsonSerializer.Deserialize<T>(json, options);
+            return JsonSerializer.Deserialize<T>(content, options);
+        }
+
+        public async Task<List<SshKey>> GetFlatSshKeysAsync(string url)
+        {
+            var response = await _httpClient.GetAsync(url);
+            if (!response.IsSuccessStatusCode)
+                return new List<SshKey>();
+
+            var json = await response.Content.ReadAsStringAsync();
+
+            using var document = JsonDocument.Parse(json);
+            var root = document.RootElement;
+
+            var result = new List<SshKey>();
+
+            if (root.TryGetProperty("keys", out var keysElement))
+            {
+                foreach (var property in keysElement.EnumerateObject())
+                {
+                    var id = property.Name;
+                    var entry = property.Value;
+
+                    var pubkey = entry.GetProperty("pubkey").GetString() ?? string.Empty;
+                    var name = entry.GetProperty("meta").GetProperty("name").GetString() ?? string.Empty;
+
+                    result.Add(new SshKey
+                    {
+                        Id = id,
+                        Pubkey = pubkey,
+                        Name = name
+                    });
+                }
+            }
+
+            return result;
+        }
+
+        public async Task<SessionSsh?> GetSessionSshAsync(string url)
+        {
+            var response = await _httpClient.GetAsync(url);
+            if (!response.IsSuccessStatusCode)
+                return null;
+
+            var json = await response.Content.ReadAsStringAsync();
+
+            using var document = JsonDocument.Parse(json);
+            var root = document.RootElement;
+
+            if (!root.TryGetProperty("ssh", out var sshElement))
+                return null;
+
+            var authorizedPubKeys = new List<string>();
+            if (sshElement.TryGetProperty("authorizedPubKeys", out var keysElement) && keysElement.ValueKind == JsonValueKind.Array)
+            {
+                foreach (var key in keysElement.EnumerateArray())
+                {
+                    authorizedPubKeys.Add(key.GetString() ?? string.Empty);
+                }
+            }
+
+            var reversePort = sshElement.GetProperty("reversePort").GetInt32();
+            var raServerUrl = sshElement.GetProperty("raServerUrl").GetString() ?? string.Empty;
+            var raServerSshPubKey = sshElement.GetProperty("raServerSshPubKey").GetString() ?? string.Empty;
+            var expiresAt = sshElement.GetProperty("expiresAt").GetDateTime();
+
+            return new SessionSsh
+            {
+                AuthorizedPubKeys = authorizedPubKeys,
+                ReversePort = reversePort,
+                RaServerUrl = raServerUrl,
+                RaServerSshPubKey = raServerSshPubKey,
+                ExpiresAt = expiresAt
+            };
         }
 
         public async Task ChangeSecretAndSave(string newSecret)
@@ -140,6 +224,82 @@ namespace ToradexSwLoader.Services
         {
             return await _httpClient.PostAsJsonAsync("https://app.torizon.io/api/v2beta/updates", deviceDto);
         }
+
+        private string FormatDuration(int durationMinutes)
+        {
+            var ts = TimeSpan.FromMinutes(durationMinutes);
+
+            if (ts.Hours > 0 && ts.Minutes > 0)
+                return $"{ts.Hours}h{ts.Minutes}m";
+            else if (ts.Hours > 0)
+                return $"{ts.Hours}h";
+            else
+                return $"{ts.Minutes}m";
+        }
+
+        public async Task<HttpResponseMessage> SendCreateSession(string deviceUuid, int durationMinutes, string? publicKey = null)
+        {
+            string sshPubKey = publicKey?.Trim() ?? "";
+
+            if (string.IsNullOrWhiteSpace(sshPubKey))
+            {
+                var baseUrl = $"https://app.torizon.io/api/v2beta/remote-access/device/{deviceUuid}/sessions";
+                var getResponse = await _httpClient.GetAsync(baseUrl).ConfigureAwait(false);
+
+                if (!getResponse.IsSuccessStatusCode)
+                    return getResponse;
+
+                var getContent = await getResponse.Content.ReadAsStringAsync().ConfigureAwait(false);
+
+                RemoteSessionInfo? sessionInfo;
+                try
+                {
+                    sessionInfo = JsonSerializer.Deserialize<RemoteSessionInfo>(getContent, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
+                }
+                catch (JsonException ex)
+                {
+                    return new HttpResponseMessage(HttpStatusCode.InternalServerError)
+                    {
+                        Content = new StringContent($"Erro ao desserializar resposta: {ex.Message}")
+                    };
+                }
+
+                sshPubKey = sessionInfo?.Ssh?.RaServerSshPubKey?.Trim() ?? "";
+            }
+
+            var payload = new
+            {
+                publicKeys = new[] { sshPubKey },
+                sessionDuration = FormatDuration(durationMinutes),
+            };
+
+            var json = JsonSerializer.Serialize(payload);
+            var content = new StringContent(json, Encoding.UTF8, "application/json");
+
+            var postUrl = $"https://app.torizon.io/api/v2beta/remote-access/device/{deviceUuid}/sessions";
+            var postResponse = await _httpClient.PostAsync(postUrl, content).ConfigureAwait(false);
+
+            _ = Task.Run(async () =>
+            {
+                await Task.Delay(TimeSpan.FromMinutes(durationMinutes));
+                await CancelSessionAsync(deviceUuid);
+            });
+
+            return postResponse;
+        }
+
+        public async Task CancelSessionAsync(string deviceUuid)
+        {
+            var deleteUrl = $"https://app.torizon.io/api/v2beta/remote-access/device/{deviceUuid}/sessions";
+            var deleteResponse = await _httpClient.DeleteAsync(deleteUrl).ConfigureAwait(false);
+
+            if (!deleteResponse.IsSuccessStatusCode)
+            {
+                var error = await deleteResponse.Content.ReadAsStringAsync();
+                Console.WriteLine($"Erro ao cancelar sessão: {deleteResponse.StatusCode} - {error}");
+            }
+        }
+
 
         public async Task<HttpResponseMessage> SendCancelAsync(List<string> deviceUuid)
         {
